+++ packages/amesos/src/SuiteSparse/KLU/Source/amesos_klu_impl.c
@@ -116,11 +116,11 @@ size_t KLU_kernel_factor	    /* 0 if failure, size of LU if OK */
     {
 	Lsize = -Lsize ;
 	Lsize = MAX (Lsize, 1.0) ;
-	lsize = Lsize * anz + n ;
+	lsize = (int) ( Lsize * anz + n ) ;
     }
     else
     {
-	lsize = Lsize ;
+	lsize = (int) Lsize ;
     }
 
     usize = lsize ;
@@ -130,8 +130,8 @@ size_t KLU_kernel_factor	    /* 0 if failure, size of LU if OK */
 
     maxlnz = (((double) n) * ((double) n) + ((double) n)) / 2. ;
     maxlnz = MIN (maxlnz, ((double) INT_MAX)) ;
-    lsize  = MIN (maxlnz, lsize) ;
-    usize  = MIN (maxlnz, usize) ;
+    lsize  = (int) ( MIN (maxlnz, lsize) ) ;
+    usize  = (int) ( MIN (maxlnz, usize) ) ;
 
     PRINTF (("Welcome to klu: n %d anz %d k1 %d lsize %d usize %d maxlnz %g\n",
 	n, anz, k1, lsize, usize, maxlnz)) ;
